requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
filelock==3.13.4
fonttools==4.51.0
fsspec==2024.3.1
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.3
joblib==1.4.0
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.3.0
psutil==5.9.8
py-cpuinfo==9.0.0
pyparsing==3.1.2
Pypubsub==4.0.3
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
schedula==1.5.10
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
sympy==1.12
tbb==2021.12.0
thop==0.1.1.post2209072238
threadpoolctl==3.4.0
torch==2.3.0+cu121
torchaudio==2.3.0+cu121
torchvision==0.18.0+cu121
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.2.2
urllib3==2.2.1
